Add features: SSH Password Auth and Oneshot Run
SSH Password Auth:
Allow the login via password with ssh for machines where no key can be
deployed. This can be configiured in the plan and you need to turn it on
in the config. This uses the sshpass tool which has to be installed.

Oneshot Run:
This allows to run a plan directly from a file which then gets removed
again at the end. This should make testing of plans easier.